- Fikset feil med Email bekreftelse - Fjernet klasser som ikke brukes i DbLib - Flyttet metode InputRecordInOrders fra klasse DbLib til klasse Register
