+++ Roombooking/src/classes/DbLib.java
@@ -10,7 +10,8 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 /**
- * DbLib provides the functionality related to the Customer entity.
+ * DbLib provides the functionality for the roombooking project.
+ * The methods used are primarily intended to update and read the database.
  * @author Karl Martin Lund
  * @version 1.0
  * @since
@@ -20,33 +21,10 @@ public class DbLib {
     private PrintWriter out;
     private Connection con;
 
-
-
-
     public DbLib(PrintWriter out) {
         this.out = out;
     }
-    /**
-     * checkForCustomer checks if there's a customer registered by either the name, email or phone submitted
-     * and maps the result by customer id and the data connected with it in a hashmap.
-     * @param name
-     * @param email
-     * @param phone
-     * @return a hashmap with the search results.
-     */
-    public HashMap<String, String> checkForCustomer(String name, String email, String phone) throws SQLException {
-        String cus_id;
-        String[] columnArray = {"cus_name", "cus_email", "cus_phone"};
-        String[] searchValueArray = {name, email, phone};
-        HashMap<String, String> searchResults = new HashMap();
-        for(int i = 0; i < columnArray.length; i++){
-            cus_id = getField("cus_id", "Customer", columnArray[i], searchValueArray[i]);
-            if(cus_id != null){
-                searchResults.put(searchValueArray[i], cus_id);
-            }
-        }
-        return searchResults;
-    }
+
     /**
      * getField will take parameters to create a SQL-statement that retrieves any single record of the
      * RoombookingDB-schema.
@@ -80,96 +58,6 @@ public class DbLib {
             if (con != null){
                 con.close();
             }
-
-
-        }
-        return null;
-    }
-
-    /**
-     * inputRecordInCustomer will create a new record of a customer in table Customer.
-     * PS: Sindre har laget en form for denne i Register.java, men jeg trengte cus_id i kalleren (for å kunne
-     * registrere kunde før ordre), så enten så må generateId skje i kalleren (createBooking) eller så må
-     * registerCustomer returnere cus_id'en. Ville ikke endre Sindre sin klasse så jeg lagde min egen her. Dvs. at
-     * jeg per nå kaller generateId i createBooking.
-     * @param cus_id
-     * @param name
-     * @param phone
-     * @param email
-     * @param password
-     * @param bonuspoints
-     */
-    public void inputRecordInCustomer(String cus_id, String name, String phone, String email, String password, String bonuspoints) throws SQLException {
-        con = new DbTool().logIn(out);
-        try {
-            PreparedStatement pst = con.prepareStatement("INSERT INTO RoombookingDB.Customer VALUES (?, ?, ?, ?, ?, ?)");
-            pst.setString(1, cus_id);
-            pst.setString(2, name);
-            pst.setString(3, phone);
-            pst.setString(4, email);
-            pst.setString(5, password);
-            pst.setString(6, bonuspoints);
-
-            int customerCount = pst.executeUpdate();
-            if(customerCount == 1) {
-                out.println(customerCount + " customer registered.");
-            }
-            else {
-                out.println(customerCount + " customer registered.");
-            }
-        }
-        catch (SQLException e) {
-            out.println("Exeption in inputRecordInCustomer: " + e);
-        } finally {
-            if (out != null) {
-               out.close();
-            }
-            if (con != null){
-                con.close();
-            }
-
-        }
-    }
-    /**
-     * inputRecordInOrders will put in a full record in table Orders. This is essentially the last part of creating
-     * an order this will stay private as it needs only be accessed from createOrder().
-     * @param room_id the id of the room that is ordered
-     * @param cus_id the id of the customer that is ordering.
-     * @param checkInDate the check in-date of the order.
-     * @param checkOutDate the check out-date of the order.
-     * @param customerPreferences any special preferences the customer has submitted.
-     * @return the order_id of the order recorded.
-     */
-    public String inputRecordInOrders(String room_id, String cus_id, String checkInDate, String checkOutDate, String customerPreferences, String paymentType) throws SQLException {
-        con = new DbTool().logIn(out);
-        try {
-            java.util.Date checkInDatePR = new SimpleDateFormat("yyyy-MM-dd").parse(checkInDate);
-            java.util.Date checkOutDatePR = new SimpleDateFormat("yyyy-MM-dd").parse(checkOutDate);
-            GenerateID idGenerator = new GenerateID();
-            String orderId = idGenerator.getID(out, "SELECT count(*) FROM RoombookingDB.Orders");
-
-            PreparedStatement pst = con.prepareStatement("INSERT INTO RoombookingDB.Orders VALUES (?, ?, ?, ?, ?, ?, ?)");
-            pst.setString(1, orderId);
-            pst.setString(2, room_id);
-            pst.setString(3, cus_id);
-            pst.setDate(4, new java.sql.Date(checkInDatePR.getTime()));
-            pst.setDate(5, new java.sql.Date(checkOutDatePR.getTime()));
-            pst.setString(6, customerPreferences);
-            pst.setString(7, paymentType);
-
-            int orderCount = pst.executeUpdate();
-            if(orderCount == 1) {
-                //out.println(orderCount + " order created.");
-            }
-            else {
-               // out.println(orderCount + " orders created.");
-            }
-            return orderId;
-        }
-        catch (SQLException | ParseException e) {
-            out.println("Exeption in createOrder: " + e);
-        } finally {
-            con.close();
         }
         return null;
     }
