+++ Roombooking/src/servlets/userServlets/CreateUserServlet.java
@@ -67,11 +67,8 @@ public class CreateUserServlet extends HttpServlet {
             String subject = "Created user in Cohesion Hotel";
             String text = "Congratulations with your new user! \n Now you can log in with your Email and Password.";
             try {
-                System.out.println("før sendEmail");
                 Email.sendMail(email, subject, text);
-                System.out.println("etter SendMail");
             } catch (Exception ex) {
-                System.out.println("i catch");
                 ex.printStackTrace();
             }
         }
